+++ SQLite-Net.Extensions.Readers/SQLiteCommandReaderExtensions.cs
@@ -13,7 +13,7 @@ namespace SQLite_Net.Extensions.Readers
     public static class SQLiteCommandReaderExtensions
     {
         internal const string DateTimeExactStoreFormat = "yyyy'-'MM'-'dd'T'HH':'mm':'ss'.'fff";
-
+        
         /// <summary>
         /// Execute a query on the database
         /// </summary>
@@ -202,7 +202,7 @@ namespace SQLite_Net.Extensions.Readers
                 case SQLite3.ColType.Float:
                     return nullable ? typeof(double?) : typeof(double);
                 case SQLite3.ColType.Integer:
-                    return nullable ? typeof(int?) : typeof(int);
+                    return nullable ? typeof(long?) : typeof(long); 
                 default:
                     return typeof(string);
             }
